fix: errors from mypy

+++ qolmat/imputations/diffusions/utils.py
@@ -2,7 +2,7 @@ import numpy as np
 import torch
 
 
-def get_num_params(model: torch.nn.Module) -> float:
+def get_num_params(model: torch.nn.Module) -> int:
     """Get the total number of parameters of a model
 
     Parameters
@@ -18,4 +18,4 @@ def get_num_params(model: torch.nn.Module) -> float:
     model_parameters = filter(lambda p: p.requires_grad, model.parameters())
     params = sum([np.prod(p.size()) for p in model_parameters])
 
-    return params
+    return int(params)
